#16 바텀 네비 구현
