+++ app/src/main/java/com/chapter/chapterkeep/ui/navigate/NavGraph.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.chapter.chapterkeep.ui.screen.boardScreen.BaekiljangScreen
+import com.chapter.chapterkeep.ui.screen.boardScreen.BaekiljangScreen.BaekiljangScreen
 import com.chapter.chapterkeep.ui.screen.boardScreen.BoardScreen
 import com.chapter.chapterkeep.ui.screen.bookScreen.AddBookScreen
 import com.chapter.chapterkeep.ui.screen.bookScreen.ViewBookScreen
